requirements.txt
Jinja2~=3.1.6
openai~=1.72.0
langchain~=0.3.23
langchain-community~=0.3.21
langchain-core~=0.3.51
scikit-learn~=1.6.1
pandas~=2.2.3
nltk~=3.9.1
filetype~=1.2.0
PyPDF2~=3.0.1
tabula-py
beautifulsoup4
python-docx
bs4~=0.0.2
py2neo~=2021.2.4
sklearn-crfsuite~=0.5.0
transformers~=4.51.2
neo4j~=5.18.0